SimulationFramework:  * Fixed rounding bug (due to incorrect cast) that was preventing correct detection of current position, and creating problems for area coverage algorithms.  * Diagonal-movement issue still present (but unrelated to port from lua to C++).

+++ DroneController/packages/madara_client/utilities/gps_utils.cpp
@@ -56,7 +56,7 @@ double SMASH::Utilities::gps_coordinates_distance (double lat1, double long1,
 // Calculates the latitude and longitude of a point given as a distance in 
 // meters in x and y from a reference point.
 ///////////////////////////////////////////////////////////////////////////////
-SMASH::Utilities::Position SMASH::Utilities::getLatAndLong(int x, int y, 
+SMASH::Utilities::Position SMASH::Utilities::getLatAndLong(double x, double y, 
   SMASH::Utilities::Position referencePoint)
 {
   // Get the deltas from reference point.
